backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
sqlmodel
pydantic
pydantic[email]
python-jose[cryptography]
python-multipart
axios
passlib==1.7.4
bcrypt==4.0.1